Combine two XCTestCase extensions
